+++ src/components/src/main/ToDoDashboad.tsx
@@ -1,3 +1,5 @@
+"use client";
+import * as React from "react";
 import { Button } from "@/components/ui/button";
 import {
   Card,
@@ -27,12 +29,46 @@ export interface Task {
 }
 
 export default function ToDoListsDashboard() {
+  const [taskList, setTaskList] = useState<Task[]>([]);
+  const [date, setDate] = useState<Date>(new Date());
+  const AddTask = (inputTitle: string) => {
+    setTaskList((prev) => [
+      ...prev,
+      {
+        id: (prev?.length + 1).toString(),
+        title: inputTitle,
+        isComplete: false,
+        createdAt: date.toLocaleDateString("ko-KR", {
+          month: "short",
+          day: "2-digit",
+        }),
+      },
+    ]);
+  };
+
+  const DeleteTask = (id: string) => {
+    setTaskList((prev) => prev.filter((task) => task.id !== id));
+  };
+
+  const getRelativeDate = (date?: Date) => {
+    if (!date) return "Today";
+    const today = new Date();
+    const diffTime = date.getTime() - today.getTime();
+    const diffDays = Math.ceil(diffTime / (1000 * 60 * 60 * 24));
+
+    if (diffDays === 0) return "Today";
+    if (diffDays === 1) return "Tomorrow";
+    if (diffDays === -1) return "Yesterday";
+    if (diffDays > 0) return `In ${diffDays} days`;
+    return `${Math.abs(diffDays)} days ago`;
+  };
+
   return (
     <Card className="w-full max-w-sm">
       <CardHeader>
-        <CardTitle>{상대날짜 ? 상대날짜 : "Today"}</CardTitle>
+        <CardTitle>{getRelativeDate(date)}</CardTitle>
         <CardDescription>
-          <DatePicker />
+          <DatePicker date={date} setDate={setDate} />
         </CardDescription>
       </CardHeader>
       <CardContent className="flex flex-col gap-2">
